Implement calculation of width requirements
In order to calculate the width requirements, the products are retrieved
from the database. All data required to calculate the width is retrieved
in a single query, after adding the order to the database.

The classes to create and retrieve data from an order are refactored,
in order to have the required properties for each operation, instead of
having a generic class that works for both cases.